+++ app/controllers/datasets_controller.rb
@@ -1,9 +1,9 @@
 class DatasetsController < ApplicationController
   before_action :authenticate_user_from_token!, only: [ :manifest, :files ]
-  before_action :authenticate_user!,        only: [ :new, :edit, :create, :update, :destroy, :audits, :requests, :request_access, :set_access, :download_covariates, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
+  before_action :authenticate_user!,        only: [ :new, :edit, :create, :update, :destroy, :audits, :requests, :request_access, :set_access, :create_access, :download_covariates, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
   before_action :check_system_admin,        only: [ :new, :create, :destroy, :pull_changes, :sync ]
   before_action :set_viewable_dataset,      only: [ :show, :manifest, :logo, :images, :variable_chart, :files, :pages, :request_access, :search, :add_variable_to_list, :remove_variable_from_list, :download_covariates ]
-  before_action :set_editable_dataset,      only: [ :edit, :update, :destroy, :audits, :requests, :set_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
+  before_action :set_editable_dataset,      only: [ :edit, :update, :destroy, :audits, :requests, :set_access, :create_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
   before_action :redirect_without_dataset,  only: [ :show, :manifest, :logo, :images, :variable_chart, :files, :pages, :request_access, :search, :add_variable_to_list, :remove_variable_from_list, :download_covariates, :edit, :update, :destroy, :audits, :requests, :set_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
 
   # Concerns
@@ -81,7 +81,12 @@ class DatasetsController < ApplicationController
     if @dataset_user = @dataset.dataset_users.find_by_id(params[:dataset_user_id])
       @dataset_user.update( editor: params[:editor], approved: params[:approved] )
     end
-    redirect_to requests_dataset_path(@dataset)
+    redirect_to requests_dataset_path(@dataset, dataset_user_id: @dataset_user ? @dataset_user.id : nil)
+  end
+
+  def create_access
+    @dataset_user = @dataset.dataset_users.where( user_id: params[:user_id] ).first_or_create
+    redirect_to requests_dataset_path(@dataset, dataset_user_id: @dataset_user ? @dataset_user.id : nil)
   end
 
   # GET /datasets/1/file_audits