+++ config/routes.rb
@@ -15,6 +15,7 @@ WwwSleepdataOrg::Application.routes.draw do
       get :audits
       get :request_access
       get :requests
+      post :create_access
       patch :set_access
       get "(/a/:auth_token)/manifest(/*path)", action: 'manifest', as: :manifest, format: false
       get "files((/a/:auth_token)(/m/:medium)/*path)", action: 'files', as: :files, format: false